requirements/dev.txt
asgiref==3.7.2
click==8.1.7
colorama==0.4.6
Django==4.2.6
django-cors-headers==4.3.0
django-ninja==0.22.2
gunicorn==21.2.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
psycopg==3.1.12
psycopg-binary==3.1.12
pydantic==1.10.13
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3

black==23.10.0
